+++ mongodbatlas/organizations_test.go
@@ -22,10 +22,7 @@ import (
 	"github.com/go-test/deep"
 )
 
-const (
-	orgID        = "5a0a1e7e0f2912c554080adc"
-	invitationID = "1"
-)
+const orgID = "5a0a1e7e0f2912c554080adc"
 
 func TestOrganizationsServiceOp_List(t *testing.T) {
 	t.Run("default", func(t *testing.T) {
@@ -408,279 +405,5 @@ func TestOrganizations_Delete(t *testing.T) {
 	_, err := client.Organizations.Delete(ctx, orgID)
 	if err != nil {
 		t.Fatalf("Organizations.Delete returned error: %v", err)
-	}
-}
-
-func TestOrganizations_ListUnacceptedInvitations(t *testing.T) {
-	client, mux, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc(fmt.Sprintf("/api/atlas/v1.0/orgs/%s/invites", orgID), func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, http.MethodGet)
-		_, _ = fmt.Fprint(w, `[
-               {
-               "createdAt": "2021-02-18T21:05:40Z",
-			   "expiresAt": "2021-03-20T21:05:40Z",
-			   "id": "5a0a1e7e0f2912c554080adc",
-			   "inviterUsername": "admin@example.com",
-			   "orgId": "5df7a168f10fab3a149357fb",
-			   "orgName": "jww-12-16",
-			   "roles": [
-				   "ORG_OWNER"
-			   ],
-			   "username": "wyatt.smith@example.com"},
-			   {"createdAt": "2021-02-18T21:05:40Z",
-			   "expiresAt": "2021-03-20T21:05:40Z",
-			   "id": "5a0a1e7e0f2912c554080adc",
-			   "inviterUsername": "admin@example.com",
-			   "orgId": "5df7a168f10fab3a149357fb",
-			   "orgName": "jww-12-16",
-			   "roles": [
-				   "ORG_OWNER"
-			   ],
-			   "teamIds": ["2"],
-			   "username": "wyatt.smith@example.com"}]`)
-	})
-
-	invitation, _, err := client.Organizations.ListUnacceptedInvitations(ctx, orgID, nil)
-	if err != nil {
-		t.Fatalf("Organizations.GetUnacceptedInvitation returned error: %v", err)
-	}
-
-	expected := &[]Invitation{
-		{
-			ID:              "5a0a1e7e0f2912c554080adc",
-			OrgID:           "5df7a168f10fab3a149357fb",
-			OrgName:         "jww-12-16",
-			CreatedAt:       "2021-02-18T21:05:40Z",
-			ExpiresAt:       "2021-03-20T21:05:40Z",
-			InviterUserName: "admin@example.com",
-			Username:        "wyatt.smith@example.com",
-			Roles:           []string{"ORG_OWNER"},
-		},
-		{
-			ID:              "5a0a1e7e0f2912c554080adc",
-			OrgID:           "5df7a168f10fab3a149357fb",
-			OrgName:         "jww-12-16",
-			CreatedAt:       "2021-02-18T21:05:40Z",
-			ExpiresAt:       "2021-03-20T21:05:40Z",
-			InviterUserName: "admin@example.com",
-			Username:        "wyatt.smith@example.com",
-			Roles:           []string{"ORG_OWNER"},
-			TeamIDs:         []string{"2"},
-		},
-	}
-
-	if diff := deep.Equal(invitation, expected); diff != nil {
-		t.Error(diff)
-	}
-}
-
-func TestOrganizations_GetUnacceptedInvitation(t *testing.T) {
-	client, mux, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc(fmt.Sprintf("/api/atlas/v1.0/orgs/%s/invites/%s", orgID, invitationID), func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, http.MethodGet)
-		_, _ = fmt.Fprint(w, `{
-			   "createdAt": "2021-02-18T21:05:40Z",
-			   "expiresAt": "2021-03-20T21:05:40Z",
-			   "id": "5a0a1e7e0f2912c554080adc",
-			   "inviterUsername": "admin@example.com",
-			   "orgId": "5df7a168f10fab3a149357fb",
-			   "orgName": "jww-12-16",
-			   "roles": [
-				   "ORG_OWNER"
-			   ],
-			   "username": "wyatt.smith@example.com"
-		}`)
-	})
-
-	invitation, _, err := client.Organizations.GetUnacceptedInvitation(ctx, orgID, invitationID)
-	if err != nil {
-		t.Fatalf("Organizations.GetUnacceptedInvitation returned error: %v", err)
-	}
-
-	expected := &Invitation{
-		ID:              "5a0a1e7e0f2912c554080adc",
-		OrgID:           "5df7a168f10fab3a149357fb",
-		OrgName:         "jww-12-16",
-		CreatedAt:       "2021-02-18T21:05:40Z",
-		ExpiresAt:       "2021-03-20T21:05:40Z",
-		InviterUserName: "admin@example.com",
-		Username:        "wyatt.smith@example.com",
-		Roles:           []string{"ORG_OWNER"},
-	}
-
-	if diff := deep.Equal(invitation, expected); diff != nil {
-		t.Error(diff)
-	}
-}
-
-func TestOrganizations_InviteUser(t *testing.T) {
-	client, mux, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc(fmt.Sprintf("/api/atlas/v1.0/orgs/%s/invites", orgID), func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, http.MethodPost)
-		_, _ = fmt.Fprint(w, `{
-			   "createdAt": "2021-02-18T21:05:40Z",
-			   "expiresAt": "2021-03-20T21:05:40Z",
-			   "id": "5a0a1e7e0f2912c554080adc",
-			   "inviterUsername": "admin@example.com",
-			   "orgId": "5df7a168f10fab3a149357fb",
-			   "orgName": "jww-12-16",
-			   "roles": [
-				   "ORG_OWNER"
-			   ],
-			   "username": "wyatt.smith@example.com"
-		}`)
-	})
-
-	body := &Invitation{
-		OrgID:           orgID,
-		OrgName:         "jww-12-16",
-		CreatedAt:       "2021-02-18T21:05:40Z",
-		ExpiresAt:       "2021-03-20T21:05:40Z",
-		InviterUserName: "admin@example.com",
-		Username:        "wyatt.smith@example.com",
-		Roles:           []string{"ORG_OWNER"},
-	}
-
-	invitation, _, err := client.Organizations.InviteUser(ctx, body)
-	if err != nil {
-		t.Fatalf("Organizations.InviteUser returned error: %v", err)
-	}
-
-	expected := &Invitation{
-		ID:              "5a0a1e7e0f2912c554080adc",
-		OrgID:           "5df7a168f10fab3a149357fb",
-		OrgName:         "jww-12-16",
-		CreatedAt:       "2021-02-18T21:05:40Z",
-		ExpiresAt:       "2021-03-20T21:05:40Z",
-		InviterUserName: "admin@example.com",
-		Username:        "wyatt.smith@example.com",
-		Roles:           []string{"ORG_OWNER"},
-	}
-
-	if diff := deep.Equal(invitation, expected); diff != nil {
-		t.Error(diff)
-	}
-}
-
-func TestOrganizations_UpdateInvitation(t *testing.T) {
-	client, mux, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc(fmt.Sprintf("/api/atlas/v1.0/orgs/%s/invites", orgID), func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, http.MethodPatch)
-		_, _ = fmt.Fprint(w, `{
-			   "createdAt": "2021-02-18T21:05:40Z",
-			   "expiresAt": "2021-03-20T21:05:40Z",
-			   "id": "5a0a1e7e0f2912c554080adc",
-			   "inviterUsername": "admin@example.com",
-			   "orgId": "5df7a168f10fab3a149357fb",
-			   "orgName": "jww-12-16",
-			   "roles": [
-				   "ORG_OWNER"
-			   ],
-			   "username": "wyatt.smith@example.com"
-		}`)
-	})
-
-	body := &Invitation{
-		OrgID:           orgID,
-		OrgName:         "jww-12-16",
-		CreatedAt:       "2021-02-18T21:05:40Z",
-		ExpiresAt:       "2021-03-20T21:05:40Z",
-		InviterUserName: "admin@example.com",
-		Username:        "wyatt.smith@example.com",
-		Roles:           []string{"ORG_OWNER"},
-	}
-
-	invitation, _, err := client.Organizations.UpdateInvitation(ctx, body)
-	if err != nil {
-		t.Fatalf("Organizations.UpdateInvitation returned error: %v", err)
-	}
-
-	expected := &Invitation{
-		ID:              "5a0a1e7e0f2912c554080adc",
-		OrgID:           "5df7a168f10fab3a149357fb",
-		OrgName:         "jww-12-16",
-		CreatedAt:       "2021-02-18T21:05:40Z",
-		ExpiresAt:       "2021-03-20T21:05:40Z",
-		InviterUserName: "admin@example.com",
-		Username:        "wyatt.smith@example.com",
-		Roles:           []string{"ORG_OWNER"},
-	}
-
-	if diff := deep.Equal(invitation, expected); diff != nil {
-		t.Error(diff)
-	}
-}
-
-func TestOrganizations_UpdateInvitationByID(t *testing.T) {
-	client, mux, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc(fmt.Sprintf("/api/atlas/v1.0/orgs/%s/invites/%s", orgID, invitationID), func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, http.MethodPatch)
-		_, _ = fmt.Fprint(w, `{
-			   "createdAt": "2021-02-18T21:05:40Z",
-			   "expiresAt": "2021-03-20T21:05:40Z",
-			   "id": "5a0a1e7e0f2912c554080adc",
-			   "inviterUsername": "admin@example.com",
-			   "orgId": "5df7a168f10fab3a149357fb",
-			   "orgName": "jww-12-16",
-			   "roles": [
-				   "ORG_OWNER"
-			   ],
-			   "username": "wyatt.smith@example.com"
-		}`)
-	})
-
-	body := &Invitation{
-		OrgID:           orgID,
-		ID:              invitationID,
-		OrgName:         "jww-12-16",
-		CreatedAt:       "2021-02-18T21:05:40Z",
-		ExpiresAt:       "2021-03-20T21:05:40Z",
-		InviterUserName: "admin@example.com",
-		Username:        "wyatt.smith@example.com",
-		Roles:           []string{"ORG_OWNER"},
-	}
-
-	invitation, _, err := client.Organizations.UpdateInvitation(ctx, body)
-	if err != nil {
-		t.Fatalf("Organizations.UpdateInvitation returned error: %v", err)
-	}
-
-	expected := &Invitation{
-		ID:              "5a0a1e7e0f2912c554080adc",
-		OrgID:           "5df7a168f10fab3a149357fb",
-		OrgName:         "jww-12-16",
-		CreatedAt:       "2021-02-18T21:05:40Z",
-		ExpiresAt:       "2021-03-20T21:05:40Z",
-		InviterUserName: "admin@example.com",
-		Username:        "wyatt.smith@example.com",
-		Roles:           []string{"ORG_OWNER"},
-	}
-
-	if diff := deep.Equal(invitation, expected); diff != nil {
-		t.Error(diff)
-	}
-}
-
-func TestOrganizations_DeleteInvitation(t *testing.T) {
-	client, mux, teardown := setup()
-	defer teardown()
-
-	mux.HandleFunc(fmt.Sprintf("/api/atlas/v1.0/orgs/%s/invites/%s", orgID, invitationID), func(w http.ResponseWriter, r *http.Request) {
-		testMethod(t, r, http.MethodDelete)
-	})
-
-	_, err := client.Organizations.DeleteInvitation(ctx, orgID, invitationID)
-	if err != nil {
-		t.Fatalf("Organizations.DeleteInvitation returned error: %v", err)
 	}
 }
